implemented redis store as feature toggle. Split local.ts/server.ts into application.ts. Configuration fixes.
