+++ api/.env.defaults
@@ -1,47 +1,11 @@
-# DO NOT EDIT THIS FILE UNLESS REQUIRED! Any local changes can be put into your own .env file.
-# ALSO: Please be aware when/if committing this file, we do not commit the secrets!
-
-# SECRETS
-IDAM_SECRET=
-S2S_SECRET=
-PUI_ENV=local
-
-# API CONFIG
-SESSION_SECRET=secretSauce
-NODE_TLS_REJECT_UNAUTHORIZED=1
-NOW=false
-IDAM_CLIENT=xuimowebapp
-OAUTH_CALLBACK_URL=/oauth2/callback
-MAX_LOG_LINE=80
-EXCEPTION_OPTIONS_MAX_LINES=1
-INDEX_URL=/
-LOGGING=debug
-PROTOCOL=http
-
-# COOKIE SETTINGS
-SECURE_COOKIE=false
-COOKIE_TOKEN=__auth__
-COOKIE_USER_ID=__userid__
-MICROSERVICE=xui_webapp
-
-# SERVICE URLS
-CCD_DATA_API_SERVICE=http://ccd-data-store-api-demo.service.core-compute-demo.internal
-CCD_DEFINITION_API_SERVICE=http://ccd-definition-store-api-demo.service.core-compute-demo.internal
-IDAM_API_SERVICE=https://idam-api.aat.platform.hmcts.net
-IDAM_WEB_SERVICE=https://idam-web-public.aat.platform.hmcts.net
-RD_PROFESSIONAL_API_SERVICE=http://rd-professional-api-aat.service.core-compute-aat.internal
-S2S_SERVICE=http://rpe-service-auth-provider-aat.service.core-compute-aat.internal
-FEE_AND_PAY_API=http://payment-api-aat.service.core-compute-aat.internal
-SERVICES_TERMS_AND_CONDITIONS_API_SERVICE=http://xui-terms-and-conditions-aat.service.core-compute-aat.internal
-
-# FEATURES
-FEATURE_PROXY_ENABLED=true
-FEATURE_TERMS_AND_CONDITIONS_ENABLED=false
+NODE_CONFIG_DIR=../config
+ALLOW_CONFIG_MUTATIONS=1
+NODE_CONFIG_ENV=development
 
 # PROXY
-PROXY_HOST=172.16.0.7
-PROXY_PORT=8080
+ROARR_LOG=false # enable this for more verbose logging when going through proxy
+MO_HTTP_PROXY=http://172.16.0.7:8080
+MO_NO_PROXY=localhost
 
-# OTHER
-APPINSIGHTS_INSTRUMENTATIONKEY=43110
-
+## PLEASE NOTE: NO FURTHER CHANGES SHOULD GO INTO THIS FILE, AS IT'S USED IN THE TEST PIPELINE
+## IF CHANGES ARE REQUIRED LOCALLY, PLEASE CREATE A "local-development.json" FILE IN THE CONFIG FOLDER
