--- api/.env.example
@@ -1,3 +1,0 @@
-# SECRETS
-IDAM_SECRET=
-S2S_SECRET=
