+++ api/configuration/index.ts
@@ -1,6 +1,5 @@
 import * as propertiesVolume from '@hmcts/properties-volume'
 import * as config from 'config'
-import { propsExist } from '../lib/objectUtilities'
 import {DEVELOPMENT, HTTP} from './constants'
 import {ENVIRONMENT, PROTOCOL} from './references'
 
@@ -18,6 +17,11 @@ export const initialiseSecrets = () => {
   propertiesVolume.addTo(config)
   // propertiesVolume.addTo(config, { mountPoint: '/Volumes/mnt/secrets/'})
 }
+
+/**
+ * Allows us to integrate the Azure key-vault flex volume, so that we are able to access Node configuration values.
+ */
+initialiseSecrets()
 
 /**
  * Get Environment
