+++ charts/xui-mo-webapp/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 name: xui-mo-webapp
 home: https://github.com/hmcts/rpx-xui-manage-organisations
-version: 0.1.35
+version: 0.1.39
 description: Expert UI
 maintainers:
   - name: HMCTS RPX XUI
