+++ charts/xui-mo-webapp/requirements.yaml
@@ -2,3 +2,7 @@ dependencies:
   - name: nodejs
     version: 1.9.0
     repository: '@hmctspublic'
+  - name: redis
+    version: ~7.1.0
+    repository: "@stable"
+    condition: redis.enabled
