+++ charts/xui-mo-webapp/values.yaml
@@ -1,3 +1,13 @@
+redis:
+  enabled: false
+  cluster:
+    enabled: false
+    slaveCount: 0
+  password: fake-password
+  master:
+    persistence:
+      enabled: false
+
 nodejs:
   aadIdentityName: xui
   applicationPort: 3000
@@ -11,8 +21,8 @@ nodejs:
     PUI_ENV: prod
     SESSION_SECRET: secretSauce
     NODE_TLS_REJECT_UNAUTHORIZED: 1
-    PROXY_HOST: 172.16.0.7
-    PROXY_PORT: 0
+    MO_HTTP_PROXY: http://172.16.0.7:8080
+    MO_NO_PROXY: localhost
     PORT: 3000
     NOW: false
     IDAM_CLIENT: xuimowebapp
@@ -35,6 +45,15 @@ nodejs:
     FEATURE_PROXY_ENABLED: false
     FEATURE_TERMS_AND_CONDITIONS_ENABLED: false
     FEATURE_HELMET_ENABLED: true
+    FEATURE_REDIS_ENABLED: true
+
+    # Redis
+    REDIS_HOST: xui-webapp-{{ .Values.global.environment }}.redis.cache.windows.net
+    REDIS_PORT: 6380
+    REDIS_SSL_ENABLED: true
+    REDIS_KEY_PREFIX: 'activity:'
+    REDIS_ACTIVITY_TTL: 5
+    REDIS_USER_DETAILS_TTL: 6000
 
     # IDAM_URL: https://idam-api.{{ .Values.global.environment }}.platform.hmcts.net
     # S2S_URL: http://rpe-service-auth-provider-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal
