+++ config/custom-environment-variables.json
@@ -25,10 +25,6 @@
         "termsAndConditions": "SERVICES_TERMS_AND_CONDITIONS_API_SERVICE"
     },
     "sessionSecret": "SESSION_SECRET",
-    "proxyConfig": {
-        "host": "PROXY_HOST",
-        "port": "PROXY_PORT"
-    },
     "secrets": {
       "rpx": {
         "AppInsightsInstrumentationKey": "APPINSIGHTS_INSTRUMENTATIONKEY",
@@ -56,6 +52,28 @@
       "helmetEnabled": {
         "__name": "FEATURE_HELMET_ENABLED",
         "__format": "json"
+      },
+      "redisEnabled": {
+        "__name": "FEATURE_REDIS_ENABLED",
+        "__format": "json"
       }
+    },
+    "helmet": {
+      "__name": "HELMET_CONFIG",
+      "__format": "json"
+    },
+    "redis": {
+      "host": "REDIS_HOST",
+      "port": {
+        "__name": "REDIS_PORT",
+        "__format": "json"
+      },
+      "tls": {
+        "__name": "REDIS_SSL_ENABLED",
+        "__format": "json"
+      },
+      "prefix": "REDIS_KEY_PREFIX",
+      "ttl": "REDIS_USER_DETAILS_TTL",
+      "password": "REDIS_PASSWORD"
     }
 }
