+++ infrastructure/main.tf
@@ -60,6 +60,11 @@ module "app" {
         FEATURE_PROXY_ENABLED = "${var.feature_proxy_enabled}"
         FEATURE_TERMS_AND_CONDITIONS_ENABLED = "${var.feature_terms_and_conditions_enabled}"
         FEATURE_HELMET_ENABLED = "${var.feature_helmet_enabled}"
+        FEATURE_REDIS_ENABLED = "${var.feature_redis_enabled}"
+
+        // Redis Cloud
+        REDISCLOUD_URL = "redis://ignore:${urlencode(module.redis-cache.access_key)}@${module.redis-cache.host_name}:${module.redis-cache.redis_port}?tls=true"
+        REDIS_ENCRYPTION_SECRET = "${module.redis-cache.access_key}"
 
         # COOKIE SETTINGS
         COOKIE_TOKEN = "${var.cookie_token}"
@@ -76,6 +81,10 @@ module "app" {
         FEE_AND_PAY_API = "${var.fee_and_pay_api}"
 
         TERRAFORM_TEST = "${var.terraform_test}"
+
+        # PROXY (If required)
+        AO_HTTP_PROXY = "${var.mo_http_proxy}"
+        AO_NO_PROXY = "${var.mo_no_proxy}"
     }
 }
 
@@ -97,4 +106,31 @@ data "azurerm_key_vault_secret" "oauth2_secret" {
 
 provider "azurerm" {
     version = "1.22.1"
+}
+
+data "azurerm_subnet" "core_infra_redis_subnet" {
+  name                 = "core-infra-subnet-1-${var.env}"
+  virtual_network_name = "core-infra-vnet-${var.env}"
+  resource_group_name  = "core-infra-${var.env}"
+}
+
+resource "azurerm_key_vault_secret" "redis_access_key" {
+  name         = "${var.product}-redis-access-key"
+  value        = "${module.redis-cache.access_key}"
+  key_vault_id = "${data.azurerm_key_vault.key_vault.id}"
+}
+
+resource "azurerm_key_vault_secret" "redis_connection_string" {
+  name = "${var.component}-redis-connection-string"
+  value = "redis://ignore:${urlencode(module.redis-cache.access_key)}@${module.redis-cache.host_name}:${module.redis-cache.redis_port}?tls=true"
+  key_vault_id = "${data.azurerm_key_vault.key_vault.id}"
+}
+
+module "redis-cache" {
+  source      = "git@github.com:hmcts/cnp-module-redis?ref=master"
+  product     = "${var.shared_product_name}-redis"
+  location    = "${var.location}"
+  env         = "${var.env}"
+  subnetid    = "${data.azurerm_subnet.core_infra_redis_subnet.id}"
+  common_tags = "${var.common_tags}"
 }
