+++ package.json
@@ -73,6 +73,7 @@
     "base-64": "^0.1.0",
     "body-parser": "^1.19.0",
     "config": "^3.2.4",
+    "connect-redis": "^4.0.4",
     "cookie-parser": "^1.4.4",
     "core-js": "^2.5.4",
     "crypto-js": "^3.1.9-1",
@@ -95,6 +96,7 @@
     "node-fetch": "^2.6.0",
     "otp": "^0.1.3",
     "p-iteration": "^1.1.7",
+    "redis": "^3.0.2",
     "rxjs": "^6.0.0",
     "rxjs-compat": "^6.4.0",
     "session-file-store": "^1.2.0",
